+++ backend/src/application/AdvancePhase.ts
@@ -2,7 +2,7 @@ import { MatchRepository } from "../domain/ports/persistance/MatchRepository";
 import { RealtimePublisher } from "../domain/ports/RealtimePublisher";
 import { MatchNotFound } from "../domain/errors";
 import { MatchResponse, MatchStatus } from "../domain/entity/match";
-import { ActionResolver } from "../domain/services/resolution/ActionResolver";
+import { ActionResolver } from "../domain/services/resolution";
 
 export interface AdvancePhaseInput {
   matchId: string;
