+++ backend/src/container.ts
@@ -10,8 +10,7 @@ import { GetMatchUseCase } from "./application/GetMatch";
 import { MatchRepository } from "./domain/ports/persistance/MatchRepository";
 import { InMemoryMatchRepository } from "./infrastructure/persistence/InMemoryMatchRepository";
 import { WebSocketPublisher } from "./infrastructure/websocket/WebSocketPublisher";
-import { ActionResolver } from "./domain/services/resolution/ActionResolver";
-import { ActionResolverFactory } from "./domain/services/resolution/ActionResolverFactory";
+import { ActionResolver, ActionResolverFactory } from "./domain/services/resolution";
 
 const wsPublisher = new WebSocketPublisher();
