+++ backend/src/domain/services/match/index.ts
@@ -1,0 +1,5 @@
+export { AbilityActionFactory } from "./AbilityActionFactory";
+export { MatchSnapshotMapper } from "./MatchSnapshotMapper";
+export { MatchVoting } from "./MatchVoting";
+export { TemplateAssignmentService } from "./TemplateAssignmentService";
+export { WinConditionEvaluator } from "./WinConditionEvaluator";
