+++ backend/src/domain/services/resolution/index.ts
@@ -1,0 +1,6 @@
+export { ActionResolver } from "./ActionResolver";
+export type { ResolutionResult } from "./ActionResolver";
+export { ActionResolverFactory } from "./ActionResolverFactory";
+export type { EffectHandler } from "./EffectHandler";
+export type { EffectResult, StateChange } from "./ResolutionContext";
+export { ResolutionContext } from "./ResolutionContext";
