+++ backend/src/infrastructure/websocket/WebSocketPublisher.ts
@@ -4,7 +4,7 @@ import {
 } from "../../domain/ports/RealtimePublisher";
 import { MatchResponse } from "../../domain/entity/match";
 import { Alignment } from "../../domain/entity/template";
-import { EffectResult } from "../../domain/services/resolution/ResolutionContext";
+import { EffectResult } from "../../domain/services/resolution";
 import { getWsManager } from "./mod";
 
 export class WebSocketPublisher implements RealtimePublisher {
